+++ frontend/src/admin/container/page/Container.tsx
@@ -1,4 +1,4 @@
-import AdminHome from "../presentation/AdminHome";
+import AdminHome from "../../presentation/page/AdminHome";
 import useGenerateProps from "./useGenerateProps";
 
 
@@ -6,3 +6,5 @@ export const AdminHomeContainer = () => {
     const generatedProps = useGenerateProps();
     return <AdminHome {...generatedProps} />;
 };
+
+export default AdminHomeContainer;
