+++ frontend/src/presentation/widgets/LoginDialog/LoginDialog.stories.tsx
